Introduced package structure
The code is organised in form of a Python package.
