+++ saltefficiency/weekly/weekly_summary_plots.py
@@ -14,169 +14,16 @@ colours.
 * Need to property parse commandline options
 
 """
-global __version__
-__version__ = "1.0"
-__date__ = "20 March 2015"
-__author__ = "Paul Kotze"
-__doc__="\nSALT Observing Summary Plots Generator, version "+__version__ +'''
-
-This script uses queries generated by the report_queries.py script and plots
-the relevant data in .png format
-
-Usage: python weekly_summary_plots.py [OPTIONS]
-
-OPTIONS are as follows, arguments are compulsory for both long and short forms.
-Example formats are shown:
-
-    -h   --help                  Prints this help
-    -s   --startdate=20141220    Sets the start date for the query
-    -e   --enddate=20150214      Sets the end date for the qeury
-    -d   --date=20150215         Sets the date for the query, if this option
-                                 is used a start date is not required, but an
-                                 interval is required
-    -i   --interval=7            Set the interval, in days, to go back in
-                                 history for the query
-'''
-__usage__='''Usage: python weekly_summary_plots.py [OPTIONS]
-
-OPTIONS are as follows, arguments are compulsory for both long and short forms.
-Example formats are shown:
-
-    -h   --help                  Prints this help
-
-Specifying a date range:
-    -s   --startdate=20141220    Sets the start date for the query
-    -e   --enddate=20150214      Sets the end date for the qeury
-
-Specifying a date and a range in days to query in the past:
-    -d   --date=20150215         Sets the date for the query, if this option
-                                 is used a start date is not required, but an
-                                 interval is required
-    -i   --interval=7            Set the interval, in days, to go back in
-                                 history for the query
-'''
-
-
-
-
 import sys
 import os
-import getopt
 import pandas as pd
 import pandas.io.sql as psql
 import MySQLdb
 import matplotlib.pyplot as pl
-import report_queries as rq
+import saltefficiency.util.report_queries as rq
 import numpy as np
 import matplotlib.dates as mdates
-from datetime import datetime
 
-def usage():
-    print __usage__
-    raise SystemExit(2)
-
-def validate_date(date_text):
-    '''
-    this function validate the dates provided
-    '''
-
-    try:
-        date = datetime.strptime(date_text, '%Y%m%d')
-    except ValueError:
-        raise ValueError("Incorrect data format, date should be YYYYMMDD")
-
-    return date
-
-def validate_int(number):
-
-    try:
-        num = int(number)
-    except ValueError:
-        raise ValueError('The interval should be an integer')
-
-    return num
-
-def parse_commandline(argv):
-    # executes if module is run from the command line
-
-    # check whether a commandline has been specified
-    if len(argv)==0:
-        usage()
-        sys.exit(2)
-    else:
-        pass
-
-    # read command line options
-    try:
-        opts,args = getopt.getopt(argv,"hs:e:d:i:",
-                ["help","startdate=","enddate=", "date=","interval="])
-    except getopt.GetoptError, inst:
-        print inst
-        print 'Use --help to get a list of options'
-        sys.exit(2)
-
-    # initiate the val_flags
-    s_val = False
-    e_val = False
-    d_val = False
-    i_val = False
-
-    # parse them to the relevant variables
-    for opt, arg in opts:
-        if opt in ('--help'):
-            usage()
-        elif opt in ('-s','--startdate'):
-            s_date = validate_date(arg)
-            s_val = True
-        elif opt in ('-e','--enddate'):
-            e_date = validate_date(arg)
-            e_val = True
-        elif opt in ('-d','--date'):
-            d_date = validate_date(arg)
-            d_val = True
-        elif opt in ('-i','--interval'):
-            inter = validate_int(arg)
-            i_val = True
-        else:
-            print 'Unknown option: ' + opt
-            usage()
-
-    # check all the flags and inform the user if there are missing values
-    if s_val + e_val == 2:
-
-        # check that the end date is after the start date
-        date_diff = (e_date - s_date).days
-
-        if date_diff < 0:
-            raise ValueError('The start date cannot be later than the end date')
-        else:
-            date = datetime.strftime(e_date, '%Y-%m-%d')
-            interval = date_diff
-
-    # check that a start AND an end date is specified
-    elif s_val + e_val == 1:
-        if s_val:
-            raise ValueError('You have to specify an end date')
-        elif e_val:
-            raise ValueError('You have to specify a start date')
-        else:
-            pass
-    else:
-        pass
-
-    if d_val + i_val == 2:
-        date = datetime.strftime(d_date, '%Y-%m-%d')
-        interval = inter
-    # check that a date AND an interval is specified
-    elif d_val + i_val == 1:
-        if d_val:
-            raise ValueError('You have to specify an interval with a date')
-        elif i_val:
-            raise ValueError('You have to specify a date with an interval')
-    else:
-        pass
-
-    return date, interval
 
 def priority_breakdown_pie_chart(x, ds, dirname='./logs/'):
     '''
@@ -289,14 +136,14 @@ def weekly_time_breakdown(x, ds, dirname='./logs/'):
 
     fig = pl.figure(figsize=(10,4),facecolor='w')
     ax = fig.add_subplot(111)
-    width = 0.65
+    width = 1.0
     ax.grid(which='major', axis='y')
     # science time per day
     s = ax.bar(x['Date'],
                 x['Science'],
                 width,
                 color = 'b',
-                edgecolor='w')
+                edgecolor='none')
 
     # engineering time per day
     e = ax.bar(x['Date'],
@@ -304,7 +151,7 @@ def weekly_time_breakdown(x, ds, dirname='./logs/'):
                 width,
                 bottom = x['Science'],
                 color = 'c',
-                edgecolor='w')
+                edgecolor='none')
 
     # weather time per day
     w = ax.bar(x['Date'],
@@ -312,7 +159,7 @@ def weekly_time_breakdown(x, ds, dirname='./logs/'):
                 width,
                 bottom = x['Science'] + x['Engineering'],
                 color = 'g',
-                edgecolor='w')
+                edgecolor='none')
 
     # problem time per day
     p = ax.bar(x['Date'],
@@ -320,7 +167,7 @@ def weekly_time_breakdown(x, ds, dirname='./logs/'):
                 width,
                 bottom = x['Science'] + x['Engineering'] + x['Weather'],
                 color = 'r',
-                edgecolor='w')
+                edgecolor='none')
 
 
     ax.set_ylabel('Hours', fontsize=11)
@@ -374,11 +221,9 @@ if __name__=='__main__':
                 port=3306,user=os.environ['SDBUSER'],
                 passwd=os.environ['SDBPASS'], db='sdb')
 
-#    obsdate = sys.argv[1]
-#    date = '{}-{}-{}'.format(obsdate[0:4], obsdate[4:6], obsdate[6:8])
-#    interval = sys.argv[2]
-
-    date, interval = parse_commandline(sys.argv[1:])
+    obsdate = sys.argv[1]
+    date = '{}-{}-{}'.format(obsdate[0:4], obsdate[4:6], obsdate[6:8])
+    interval = sys.argv[2]
 
     # use the connection to get the required data: _d
     dr_d = rq.date_range(mysql_con, date, interval=interval)
